Moved different sections onto different websites
Just messing around. Currently on a different repository ofc from the original. I need to change the contact one slightly. Something's up with the bottom section. Does it have to do with the heading? Or maybe I copy and pasted wrong?
Side note: I apparently miseed a </div> so that's why. Technically a copying error.